+++ GameBackboneSln/GameBackboneExe/source/DemoRegion.cpp
@@ -9,4 +9,4 @@ void DemoRegion::handleMouseMove(sf::Vector2f mousePosition) {}
 
 void DemoRegion::handleMouseClick(sf::Vector2f newPosition, sf::Mouse::Button button){}
 
-void DemoRegion::handleWheelScroll(int scrollDelta){}
+void DemoRegion::handleWheelScroll(float scrollDelta){}
